Extract SchemaLoader from core.js
